refactor(web): use shared handleApiResponse in service files
- auth.service.ts: remove local helper, import from ~/lib
- user.service.ts: remove local helper, import from ~/lib
- vehicle.service.ts: remove local helper, import from ~/lib

Closes #9
