+++ apps/web/src/features/auth/api/auth.service.ts
@@ -1,20 +1,10 @@
-import { api } from '~/lib/apiRequest'
-import type { ApiResponse, User } from '~/types'
-
-/**
- * Helper function for API responses that return data
- */
-const handleApiResponseWithData = <T>(response: ApiResponse<T>): T => {
-  if (response.success && response.data) {
-    return response.data
-  }
-  throw new Error(response.message || 'API request failed')
-}
+import { api, handleApiResponse } from '~/lib'
+import type { User } from '~/types'
 
 /**
  * Get current authenticated user (SSR-only - use useAuth hook on client)
  */
 export const getCurrentUser = async (): Promise<User> => {
   const response = await api.get<User>('/users/me')
-  return handleApiResponseWithData(response)
+  return handleApiResponse(response)
 }
